backend/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2.1
django-multiselectfield==0.1.8
django-rest-knox==4.0.1
djangorestframework==3.9.4
more-itertools==7.0.0
pluggy==0.11.0
py==1.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
